# src/main/resources/openapi.yml
openapi: 3.0.1
info:
  title: Spring PetClinic
  description: Spring PetClinic Sample Application.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: '1.0'
servers:
  - url: http://localhost:9966/petclinic/api
security:
  - bearerAuth: []
...
paths:
...
  /auth/me:
    get:
      tags: [auth]
      operationId: getCurrentUser
      summary: Get the authenticated user's current profile and authority code
      responses:
        '200':
          description: Current account profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUser'
        '401':
          description: Authentication required.
  /appointments:
    post:
      tags: [appointments]
      operationId: addAppointment
      summary: Create a future appointment
      description: Rejects overlapping scheduled appointments for either the pet or vet. Offset-aware times are stored and returned in UTC.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentFields'
      responses:
        '201':
          description: Appointment created.
          headers:
            Location:
              schema:
                type: string
              description: URL of the created appointment.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '400':
          $ref: '#/components/responses/AppointmentInvalidRequest'
        '404':
          $ref: '#/components/responses/AppointmentNotFound'
        '409':
          $ref: '#/components/responses/AppointmentConflict'
    get:
      tags: [appointments]
      operationId: listAppointments
      summary: List appointments sorted by startTime and ID
      description: Returns an empty list when there are no matches. from/to select appointments overlapping the half-open interval [from, to). All statuses are included unless filtered.
      parameters:
        - name: petId
          in: query
          schema: {type: integer, format: int32, minimum: 1}
        - name: vetId
          in: query
          schema: {type: integer, format: int32, minimum: 1}
        - name: ownerId
          in: query
          description: Current owner of the pet.
          schema: {type: integer, format: int32, minimum: 1}
        - name: status
          in: query
          description: SCHEDULED or CANCELLED.
          schema: {type: string}
        - name: from
          in: query
          schema: {type: string, format: date-time}
        - name: to
          in: query
          schema: {type: string, format: date-time}
        - name: limit
          in: query
          schema: {type: integer, format: int32, default: 20, minimum: 1, maximum: 100}
        - name: offset
          in: query
          schema: {type: integer, format: int32, default: 0, minimum: 0}
      responses:
        '200':
          description: Matching appointments.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Appointment'
        '400':
          $ref: '#/components/responses/AppointmentInvalidRequest'
  /appointments/{appointmentId}:
    parameters:
      - $ref: '#/components/parameters/AppointmentId'
    get:
      tags: [appointments]
      operationId: getAppointment
      summary: Get an appointment
      responses:
        '200':
          $ref: '#/components/responses/AppointmentResult'
        '404':
          $ref: '#/components/responses/AppointmentNotFound'
    put:
      tags: [appointments]
      operationId: updateAppointment
      summary: Replace the details of an upcoming scheduled appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentFields'
      responses:
        '200':
          $ref: '#/components/responses/AppointmentResult'
        '400':
          $ref: '#/components/responses/AppointmentInvalidRequest'
        '404':
          $ref: '#/components/responses/AppointmentNotFound'
        '409':
          $ref: '#/components/responses/AppointmentConflict'
  /appointments/{appointmentId}/cancel:
    parameters:
      - $ref: '#/components/parameters/AppointmentId'
    post:
      tags: [appointments]
      operationId: cancelAppointment
      summary: Cancel an upcoming appointment without deleting its history
      description: Repeated cancellation returns the same cancelled appointment. Appointments that have started cannot be cancelled.
      responses:
        '200':
          $ref: '#/components/responses/AppointmentResult'
        '404':
          $ref: '#/components/responses/AppointmentNotFound'
        '409':
          $ref: '#/components/responses/AppointmentConflict'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    AppointmentId:
      name: appointmentId
      in: path
      required: true
      schema: {type: integer, format: int32, minimum: 1}
  responses:
    AppointmentResult:
      description: Appointment details.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Appointment'
    AppointmentInvalidRequest:
      description: Invalid fields, dates or query parameters.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    AppointmentNotFound:
      description: Appointment, pet or vet does not exist.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    AppointmentConflict:
      description: Overlapping booking or appointment cannot be changed or cancelled.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
  schemas:
...
    AuthUser:
      type: object
      properties:
        username:
          type: string
        authCode:
          type: string
          description: OWNER, VETS or ADMIN.
        ownerId:
          type: integer
          format: int32
          nullable: true
        vetId:
          type: integer
          format: int32
          nullable: true
      required: [username, authCode]
...
    AppointmentFields:
      type: object
      description: Editable appointment details. Times must include a timezone offset. The start must be in the future and the end must follow the start.
      properties:
        petId:
          type: integer
          format: int32
          minimum: 1
          example: 1
        vetId:
          type: integer
          format: int32
          minimum: 1
          example: 1
        startTime:
          type: string
          format: date-time
          example: '2026-09-20T10:00:00+09:00'
        endTime:
          type: string
          format: date-time
          example: '2026-09-20T10:30:00+09:00'
        reason:
          type: string
          minLength: 1
          maxLength: 255
          example: Annual vaccination
      required: [petId, vetId, startTime, endTime, reason]
    Appointment:
      description: A reservation for a pet with a vet. ownerId is derived from the pet; appointment times are returned in UTC with microsecond precision.
      allOf:
        - $ref: '#/components/schemas/AppointmentFields'
        - type: object
          properties:
            id:
              type: integer
              format: int32
              readOnly: true
            ownerId:
              type: integer
              format: int32
              readOnly: true
            status:
              type: string
              description: SCHEDULED or CANCELLED. Status is managed by the server.
              readOnly: true
              example: SCHEDULED
          required: [id, ownerId, status]
    ProblemDetail:
      title: Problem Detail
      description: The schema for all error responses.
      type: object
      properties:
        type:
          title: Type
          description: Full URL that originated the error response.
          type: string
          format: uri
          example: 'http://localhost:9966/petclinic/api/owner'
          readOnly: true
        title:
          title: Title
          description: The short error title.
          type: string
          example: NoResourceFoundException
          readOnly: true
        status:
          title: Status
          description: HTTP status code
          type: integer
          format: int32
          minimum: 400
          maximum: 600
          exclusiveMaximum: true
          example: 500
          readOnly: true
        detail:
          title: Detail
          description: The long error message.
          type: string
          example: 'No static resource api/owner.'
          readOnly: true
        timestamp:
          title: Timestamp
          description: The time the error occurred.
          type: string
          format: date-time
          example: '2024-11-23T13:59:21.3820407Z'
          readOnly: true
        schemaValidationErrors:
          title: Schema validation errors
          description: Validation errors against the OpenAPI schema.
          type: array
          items:
            $ref: '#/components/schemas/ValidationMessage'
      required:
        - type
        - title
        - status
        - detail
        - timestamp
        - schemaValidationErrors
    ValidationMessage:
      title: Validation message
      description: Messages describing a validation error.
      type: object
      properties:
        message:
          title: Message
          description: The validation message.
          type: string
          example: "[Path '/lastName'] Instance type (null) does not match any allowed primitive type (allowed: ['string'])"
          readOnly: true
      required:
        - message
      additionalProperties: true
...